Refactor: change shared-android module package name
