+++ features/home/board/src/main/kotlin/team/applemango/runnerbe/feature/home/board/MainBoardFragment.kt
@@ -30,7 +30,7 @@ import team.applemango.runnerbe.feature.home.board.component.MainBoardComposable
 import team.applemango.runnerbe.feature.home.board.mvi.MainBoardState
 import team.applemango.runnerbe.shared.domain.constant.EmptyString
 import team.applemango.runnerbe.shared.domain.extension.toMessage
-import team.applemango.runnerbe.shared.extension.toast
+import team.applemango.runnerbe.shared.android.extension.toast
 
 @AndroidEntryPoint
 class MainBoardFragment : Fragment() {
