+++ features/home/board/src/main/kotlin/team/applemango/runnerbe/feature/home/board/MainBoardViewModel.kt
@@ -38,7 +38,7 @@ import team.applemango.runnerbe.domain.runningitem.model.runningitem.RunningItem
 import team.applemango.runnerbe.domain.runningitem.usecase.LoadRunningItemsUseCase
 import team.applemango.runnerbe.domain.user.usecase.UpdateBookmarkItemUseCase
 import team.applemango.runnerbe.feature.home.board.mvi.MainBoardState
-import team.applemango.runnerbe.shared.base.BaseViewModel
+import team.applemango.runnerbe.shared.android.base.BaseViewModel
 import javax.inject.Inject
 import kotlin.time.ExperimentalTime
 import kotlin.time.measureTime
