+++ features/home/write/src/main/kotlin/team/applemango/runnerbe/feature/home/write/RunningItemWriteActivity.kt
@@ -23,10 +23,10 @@ import dagger.hilt.android.AndroidEntryPoint
 import org.orbitmvi.orbit.viewmodel.observe
 import team.applemango.runnerbe.feature.home.write.component.RunningItemWrite
 import team.applemango.runnerbe.feature.home.write.mvi.RunningItemWriteState
-import team.applemango.runnerbe.shared.base.WindowInsetsActivity
+import team.applemango.runnerbe.shared.android.base.WindowInsetsActivity
 import team.applemango.runnerbe.shared.compose.theme.GradientAsset
-import team.applemango.runnerbe.shared.extension.collectWithLifecycle
-import team.applemango.runnerbe.shared.extension.finishWithAnimation
+import team.applemango.runnerbe.shared.android.extension.collectWithLifecycle
+import team.applemango.runnerbe.shared.android.extension.finishWithAnimation
 
 @AndroidEntryPoint
 class RunningItemWriteActivity : WindowInsetsActivity() {
