+++ features/register/onboard/src/main/kotlin/team/applemango/runnerbe/feature/register/onboard/component/step/EmployeeIdVerify.kt
@@ -59,7 +59,7 @@ import team.applemango.runnerbe.shared.compose.extension.parseHtml
 import team.applemango.runnerbe.shared.compose.extension.presentationDrawableOf
 import team.applemango.runnerbe.shared.compose.theme.ColorAsset
 import team.applemango.runnerbe.shared.compose.theme.Typography
-import team.applemango.runnerbe.shared.extension.toast
+import team.applemango.runnerbe.shared.android.extension.toast
 
 private val PhotoContainerModifier = Modifier
     .fillMaxWidth()
