+++ features/register/snslogin/src/main/kotlin/team/applemango/runnerbe/feature/register/snslogin/SnsLoginActivity.kt
@@ -29,14 +29,14 @@ import team.applemango.runnerbe.feature.register.snslogin.constant.LoginState
 import team.applemango.runnerbe.feature.register.snslogin.di.module.RepositoryModule
 import team.applemango.runnerbe.feature.register.snslogin.di.module.UseCaseModule
 import team.applemango.runnerbe.feature.register.snslogin.mvi.LoginSideEffect
-import team.applemango.runnerbe.shared.base.WindowInsetsActivity
+import team.applemango.runnerbe.shared.android.base.WindowInsetsActivity
 import team.applemango.runnerbe.shared.compose.extension.verticalInsetsPadding
 import team.applemango.runnerbe.shared.compose.theme.GradientAsset
-import team.applemango.runnerbe.shared.constant.DataStoreKey
-import team.applemango.runnerbe.shared.extension.changeActivityWithAnimation
-import team.applemango.runnerbe.shared.extension.collectWithLifecycle
-import team.applemango.runnerbe.shared.extension.dataStore
-import team.applemango.runnerbe.shared.extension.launchedWhenCreated
+import team.applemango.runnerbe.shared.android.constant.DataStoreKey
+import team.applemango.runnerbe.shared.android.extension.changeActivityWithAnimation
+import team.applemango.runnerbe.shared.android.extension.collectWithLifecycle
+import team.applemango.runnerbe.shared.android.extension.dataStore
+import team.applemango.runnerbe.shared.android.extension.launchedWhenCreated
 import team.applemango.runnerbe.util.DFMOnboardActivityAlias
 import team.applemango.runnerbe.util.MainActivityAlias
