+++ features/register/snslogin/src/main/kotlin/team/applemango/runnerbe/feature/register/snslogin/SnsLoginViewModel.kt
@@ -22,7 +22,7 @@ import team.applemango.runnerbe.domain.register.runnerbe.usecase.GetNaverAccessT
 import team.applemango.runnerbe.domain.register.runnerbe.usecase.LoginUseCase
 import team.applemango.runnerbe.feature.register.snslogin.constant.LoginState
 import team.applemango.runnerbe.feature.register.snslogin.mvi.LoginSideEffect
-import team.applemango.runnerbe.shared.base.BaseViewModel
+import team.applemango.runnerbe.shared.android.base.BaseViewModel
 import team.applemango.runnerbe.shared.domain.util.requireFieldNullMessage
 
 internal class SnsLoginViewModel(
