+++ presentation/src/main/kotlin/team/applemango/runnerbe/activity/MainActivity.kt
@@ -18,7 +18,7 @@ import androidx.navigation.ui.setupWithNavController
 import dagger.hilt.android.AndroidEntryPoint
 import team.applemango.runnerbe.R
 import team.applemango.runnerbe.databinding.ActivityMainBinding
-import team.applemango.runnerbe.shared.constant.BottomNavigationBarHeight
+import team.applemango.runnerbe.shared.android.constant.BottomNavigationBarHeight
 import team.applemango.runnerbe.shared.domain.unit.dp
 import team.applemango.runnerbe.shared.domain.unit.toInt
