+++ presentation/src/main/kotlin/team/applemango/runnerbe/activity/StartActivity.kt
@@ -24,13 +24,13 @@ import io.github.jisungbin.logeukes.logeukes
 import kotlinx.coroutines.cancel
 import kotlinx.coroutines.flow.cancellable
 import kotlinx.coroutines.flow.catch
-import team.applemango.runnerbe.shared.constant.DataStoreKey
+import team.applemango.runnerbe.shared.android.constant.DataStoreKey
 import team.applemango.runnerbe.shared.domain.extension.toMessage
 import team.applemango.runnerbe.shared.domain.util.flowExceptionMessage
-import team.applemango.runnerbe.shared.extension.changeActivityWithAnimation
-import team.applemango.runnerbe.shared.extension.collectWithLifecycle
-import team.applemango.runnerbe.shared.extension.dataStore
-import team.applemango.runnerbe.shared.extension.toast
+import team.applemango.runnerbe.shared.android.extension.changeActivityWithAnimation
+import team.applemango.runnerbe.shared.android.extension.collectWithLifecycle
+import team.applemango.runnerbe.shared.android.extension.dataStore
+import team.applemango.runnerbe.shared.android.extension.toast
 import team.applemango.runnerbe.util.DFMLoginActivityAlias
 import team.applemango.runnerbe.util.DFMOnboardActivityAlias
