+++ shared/android/src/main/kotlin/team/applemango/runnerbe/shared/android/extension/activity.kt
@@ -7,7 +7,7 @@
  * Please see: https://github.com/applemango-runnerbe/RunnerBe-Android/blob/main/LICENSE.
  */
 
-package team.applemango.runnerbe.shared.extension
+package team.applemango.runnerbe.shared.android.extension
 
 import android.app.Activity
 import android.content.Intent
